+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Output is no longer colored.
 
 ### Fixed
+- [Issue 21](https://github.com/hekmekk/git-team/issues/21): The `prepare-commit-msg` hook ensures that changing the `activation-scope` does not lead to inconsistent behaviour.
 - [Issue 25](https://github.com/hekmekk/git-team/issues/25): The `disable` command now returns early without taking any action if not previously enabled, in order to prevent any unjustified modification of previously set `core.hooksPath` or `commit.template` or the file the latter points to.
 
 ## [2.0.0] - 2026-01-30